storage: move LivenessStatus out of liveness.go
This method is quit idiosyncratic. The majority of its callers used to
be buggy. This patch hides the function some by moving it out of the
storagepb/liveness.go, closer to its few remaining callers.

Release note: None
